Major refactor and code cleanup
- input now accepts Iterable[str]
- Overhauling the README and updating CHANGELOG
- major code changes under the hood but the front end use should be the
  same

+++ enter-the-dojo.sh
@@ -2,9 +2,14 @@
 # Mikey Garcia, @gikeymarcia
 # meant to be sourced before returning to development
 
-figlet "enter the dojo"
-ls -l
 
-source env/bin/activate
-python -m pip install -r ./requirements_dev.txt
-python -m pip list
+activate="env/bin/activate"
+if [ -f "$activate" ]; then
+    figlet "enter the dojo"
+    ls -l
+    source env/bin/activate
+    python -m pip install -r ./requirements_dev.txt
+    python -m pip list
+else
+    echo "Could not find virtualenv activator at $activate"
+fi
